pyproject.toml
[tool.poetry]
name = "discord bot"
version = "0.1.0"
description = "Stream audio from computer input"
authors = ["Ygor Pontelo"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pyaudio = "^0.2.13"
samplerate = "^0.1.0"
discord-py = {extras = ["voice"], version = "^2.2.3"}
numpy = "^1.24.3"
questionary = "^1.10.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.264"
mypy = "^1.2.0"
pyinstaller = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
